Filter's classes moved to their own folder
