+++ generator/entities/filter-components/block.js
