+++ generator/entities/filter-components/section.js
